+++ ui/src/flux/constants/functions.ts
@@ -463,13 +463,23 @@ export const FUNCTIONS: FluxToolbarFunction[] = [
       'https://docs.influxdata.com/flux/latest/stdlib/built-in/transformations/aggregates/covariance/',
   },
   {
-    name: 'csv.from',
+    name: 'csv.from (file)',
     args: [
       {
         name: 'file',
         desc: 'The file path of the CSV file to query.',
         type: 'String',
       },
+    ],
+    package: 'csv',
+    desc: 'Retrieves data from a comma-separated value (CSV) data source.',
+    example: 'csv.from(file: path)',
+    category: 'Inputs',
+    link: 'https://docs.influxdata.com/flux/latest/stdlib/csv/from/',
+  },
+  {
+    name: 'csv.from (csvData)',
+    args: [
       {
         name: 'csv',
         desc:
@@ -484,7 +494,7 @@ export const FUNCTIONS: FluxToolbarFunction[] = [
     link: 'https://docs.influxdata.com/flux/latest/stdlib/csv/from/',
   },
   {
-    name: 'csv.from',
+    name: 'csv.from (url)',
     args: [
       {
         name: 'url',